refactoring api upload
